django_back_end/requirements.txt
asgiref==3.4.1
bson==0.5.8
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
dataclasses==0.6
defusedxml==0.7.1
Django==3.2.9
django-cors-headers==3.10.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
stripe==2.63.0
uritemplate==4.1.1
urllib3==1.26.7
uuid==1.30
wincertstore==0.2
